chg: [core] Simple refractor with an additional "Database common" class, that factorize code of both requester and adder (template only for now, no deduplication)
